Bookstore/requirements.txt
amqp==2.6.0
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.8.0
billiard==3.6.3.0
boto3==1.15.11
botocore==1.18.11
celery==4.4.6
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
colorama==0.4.4
dataclasses==0.7
defusedxml==0.6.0
Django==2.2.15
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-phonenumber-field==4.0.0
django-storages==1.10.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.1.1
jmespath==0.10.0
kombu==4.6.11
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.9
phonenumbers==8.12.6
Pillow==7.2.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-http-client==3.2.7
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sendgrid==6.4.3
six==1.15.0
sqlparse==0.2.4
starkbank-ecdsa==1.0.0
toml==0.10.2
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.4.0
